Switch error log to http log
